Initial commit of athinput-mode
This is a major mode for editing athena input files.  It's pretty
simple: it derives from conf-mode and adds indentation and syntax
highlighting.  There's also a function to automatically add the
<par_end> block if it's missing.

The only other thing worth mentioning is that this mode leaves you
with no control over the whitespace in the file.  This may seem
annoying at first, but it serves a good purpose.  Specifically, if you
keep your configuration files under version control, you can end up
with annoying "whitespace-only" commits.  This mode aims to make those
impossible.